Refs #29906. Move new files into the common directory.
