+++ qt/widgets/CMakeLists.txt
@@ -1,6 +1,5 @@
 # Now add the packages one-by-one, building up the dependencies as we go
 add_subdirectory(common)
-add_subdirectory(fitting)
 add_subdirectory(mplcpp)
 add_subdirectory(plotting)
 add_subdirectory(instrumentview)
