+++ qt/widgets/common/inc/MantidQtWidgets/Common/FitScriptGenerator.ui
