--- qt/widgets/fitting/src/FitScriptGeneratorModel.cpp
@@ -1,17 +1,0 @@
-// Mantid Repository : https://github.com/mantidproject/mantid
-//
-// Copyright &copy; 2020 ISIS Rutherford Appleton Laboratory UKRI,
-//   NScD Oak Ridge National Laboratory, European Spallation Source,
-//   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
-// SPDX - License - Identifier: GPL - 3.0 +
-#include "FitScriptGeneratorModel.h"
-
-namespace MantidQt {
-namespace MantidWidgets {
-namespace Fitting {
-
-FitScriptGeneratorModel::FitScriptGeneratorModel() {}
-
-} // namespace Fitting
-} // namespace MantidWidgets
-} // namespace MantidQt
